apple clang int conversion and doxygen format warnings

+++ src/timetag.c
@@ -57,7 +57,7 @@ void lo_timetag_now(lo_timetag * t)
     struct timeval tv;
 
     gettimeofday(&tv, NULL);
-    t->sec = tv.tv_sec + JAN_1970;
+    t->sec = (uint32_t) (tv.tv_sec + JAN_1970);
     t->frac = tv.tv_usec * 4294.967295;
 #endif
 }
